Create cloud instances and their DNS records
Creating the DNS records right away lets us follow up with plays that
require login right away.  This *could* be accomplished with a dynamic
inventory, but I'm not sure how well a dynamic inventory would interact
with the non-dynamic inventory for physical hosts and inventory
values we need prior to instantiation.