start blog branch
Not much done here, so I'll expand on the paradigm to make the work easy
to come back to.  The premise here is that site code is maintained in a
git repository.  A post-receive hook is used to checkout the ie staging branch
to a directory that serves as the DocumentRoot for the staging Apache
VirtualHost, and so with production.  The action should be extensible
for arbitrary branches, iterating over a variable dict containing
relevant personalization for the site.